Integrated the new core geometric classes. (Vector3, Point3 etc.)

+++ GeometricalObject.hpp
@@ -7,8 +7,8 @@
 
 class GeometricalObject {
  public:
-  virtual bool hit(const Ray& ray, double tMin, double tMax, HitRecord& hitRecord) const = 0;
-  virtual bool computeBoundingBox(double t0, double t1, AABB& outputBox) const = 0;
+  virtual bool hit(const Ray& ray, float tMin, float tMax, HitRecord& hitRecord) const = 0;
+  virtual bool computeBoundingBox(float t0, float t1, AABB& outputBox) const = 0;
 };
 
 using GeoObject = GeometricalObject;
